+++ birdnet-pi/CHANGELOG.md
@@ -1,3 +1,5 @@
+- [FEAT] : publish mqtt to homeassistant if a server is found
+
 ## 0.13-61 (30-06-2024)
 - [FIX] : safeguard to avoid embedded pulseaudio interference